feat: move report to helpers
